reorganize safe hooks
