+++ extension/src/providers/index.ts
@@ -2,7 +2,7 @@ export { default as ForkProvider } from './ForkProvider'
 export { default as WrappingProvider } from './WrappingProvider'
 export { default as useWalletConnect } from './useWalletConnect'
 export { useGanacheProvider, default as ProvideGanache } from './ProvideGanache'
-export { waitForMultisigExecution } from './safe'
+export { waitForMultisigExecution } from '../safe'
 export {
   useTenderlyProvider,
   default as ProvideTenderly,
